+++ front/react/src/contents/ContentsHeader.css
@@ -44,8 +44,3 @@
     color: #f5f5f5;
     margin-right: 1rem;
 }
-
-
-.groupUsers{
-  
-}
