+++ front/react/src/contents/ContentsHeader.js
@@ -3,6 +3,7 @@ import Popup2 from "../Popup2";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faBars, faFolderPlus, faUserPlus } from "@fortawesome/free-solid-svg-icons";
 import styles from "./ContentsHeader.css";
+import GroupInUserList from "./GroupInUserList";
 
 const API_URL = "http://localhost:8080/codingvirus19";
 const API_HEADERS = {
@@ -31,7 +32,6 @@ export default class ContentsHeader extends React.Component {
     // no가 null이 아닌것 에서만 그룹나가기가 작동되도록 설정.
     
     if(this.props.groupBySidebar.no != null){
-
       // 들어가있는 그룹의 no와 name값(상훈아 name넣었다!!!)
       groupNoForDelete ={
         no: this.props.groupBySidebar.no,
@@ -99,10 +99,7 @@ export default class ContentsHeader extends React.Component {
             />
           ) : null}
         </div>
-        <div className={styles.groupUsers}>
-            <div>
-            </div>
-        </div>
+        {(this.props.groupBySidebar.no != null) ? <GroupInUserList groupInUserList={this.props.groupInUserList} users={this.props.users} /> : null}
       </div>
       <div className={styles.groupOut}>
         {this.props.groupBySidebar.no != null ?
